Refactor(signatories): Remove redundant signatory data logic
This commit removes the redundant `GetViewRecordsByOffice` method from `SignatoriesHasReferencesRepository.cs` and `ISignatoriesHasReferences.cs`, and instead uses a simplified data retrieval method in `frmSignatories.cs`.

This change centralizes user-related data access by replacing direct calls to `Helper.LoggedInUserData()` with the new `UserHelper.loggedUser` properties, ensuring consistency and cleaner code across the application. It also refactors `UsersModel` to use `int` for `RoleId` and removes an unused event handler from `ucJournals.Designer.cs`.

+++ ACC.Domain/Models/UsersModel.cs
@@ -3,7 +3,7 @@
     public class UsersModel
     {
         public int Id { get; set; }
-        public byte RoleId { get; set; }
+        public int RoleId { get; set; }
         public string UserName { get; set; }
         public string Prefix { get; set; }
         public string FirstName { get; set; }
